requirements.txt
alembic==1.7.4
aniso8601==9.0.1
apispec==5.1.1
apispec-webframeworks==0.5.2
atomicwrites==1.4.0
attrs==21.2.0
CacheControl==0.12.6
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
colorama==0.4.4
coverage==6.0.1
cryptography==35.0.0
firebase-admin==5.0.3
flasgger==0.9.5
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
gevent==21.8.0
google-api-core==2.1.0
google-api-python-client==2.24.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.41.0
gunicorn==20.1.0
httplib2==0.20.1
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.0
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
mistune==0.8.4
msgpack==1.0.2
mysqlclient==2.0.3
packaging==21.0
phonenumbers==8.12.34
pluggy==1.0.0
proto-plus==1.19.4
protobuf==3.18.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.2.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dotenv==0.19.0
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.25
SQLAlchemy-Utils==0.37.8
toml==0.10.2
tomli==1.2.1
uritemplate==3.0.1
urllib3==1.26.7
Werkzeug==2.0.2
zope.event==4.5.0
zope.interface==5.4.0
